+++ internal/service/conflicts/resolver.go
@@ -1,12 +1,60 @@
 package conflicts
 
 import (
+	"fmt"
+
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
 	"gitlab.com/gitlab-org/gitaly/internal/helper"
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
 )
 
 func (s *server) ListConflictFiles(in *pb.ListConflictFilesRequest, stream pb.ConflictsService_ListConflictFilesServer) error {
-	return helper.Unimplemented
+	ctx := stream.Context()
+
+	if err := validateListConflictFilesRequest(in); err != nil {
+		return grpc.Errorf(codes.InvalidArgument, "ListConflictFiles: %v", err)
+	}
+
+	client, err := s.ConflictsServiceClient(ctx)
+	if err != nil {
+		return err
+	}
+
+	clientCtx, err := rubyserver.SetHeaders(ctx, in.GetRepository())
+	if err != nil {
+		return err
+	}
+
+	rubyStream, err := client.ListConflictFiles(clientCtx, in)
+	if err != nil {
+		return err
+	}
+
+	return rubyserver.Proxy(func() error {
+		resp, err := rubyStream.Recv()
+		if err != nil {
+			md := rubyStream.Trailer()
+			stream.SetTrailer(md)
+			return err
+		}
+		return stream.Send(resp)
+	})
+}
+
+func validateListConflictFilesRequest(in *pb.ListConflictFilesRequest) error {
+	if in.GetRepository() == nil {
+		return fmt.Errorf("empty Repository")
+	}
+	if in.GetOurCommitOid() == "" {
+		return fmt.Errorf("empty OurCommitOid")
+	}
+	if in.GetTheirCommitOid() == "" {
+		return fmt.Errorf("empty TheirCommitOid")
+	}
+
+	return nil
 }
 
 func (s *server) ResolveConflicts(stream pb.ConflictsService_ResolveConflictsServer) error {